hostal_sanrosa/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
ConfigArgParse==1.7.1
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
faiss-cpu==1.12.0
Faker==19.6.1
fastapi==0.119.0
filelock==3.20.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
fsspec==2025.9.0
geographiclib==2.1
geopy==2.3.0
gevent==25.5.1
geventhttpclient==2.3.4
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.185.0
google-auth==2.42.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
locust==2.40.4
locust-cloud==1.26.3
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numpy==1.26.0
openai==2.6.1
packaging==25.0
pika==1.3.2
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.1
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-engineio==4.12.2
python-frontmatter==1.1.0
python-socketio==5.13.0
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.10.23
reportlab==4.4.3
requests==2.31.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==6.0
zope.interface==8.0
